final single-variable regression model
The model uses only one variable, Local Young Population, as the
predictor. Previous attempts to find other significant correlations
failed, so I had to do with population only. I estimated the error
of the total predicted numbers so we have a lower and upper bound for
these estimations. I added two .csv files with these predictions so it
should be easy to break down the data by regions etc.